feat: Add Home Assistant data table integration with submodule
