+++ .github/workflows/ci.yml
@@ -15,6 +15,8 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4.2.2
+        with:
+          submodules: recursive
       - name: Use Node.js
         uses: actions/setup-node@v4.4.0
         with:
