+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "homeassistant-frontend"]
+	path = homeassistant-frontend
+	url = https://github.com/home-assistant/frontend.git
