+++ package.json
@@ -49,8 +49,10 @@
     "@polymer/paper-tooltip": "^3.0.1",
     "@types/w3c-web-serial": "^1.0.8",
     "esptool-js": "^0.5.3",
+    "home-assistant-js-websocket": "^9.5.0",
     "improv-wifi-serial-sdk": "^2.5.0",
     "lit": "^2.8.0",
+    "memoize-one": "^6.0.0",
     "tslib": "^2.8.1",
     "yaml": "^2.7.1"
   },
